requirements.txt
android==1.0
APScheduler==3.6.3
astroid==2.4.1
audiostream==0.2
backports.zoneinfo==0.2.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cmake==3.17.2
idna==3.2
isort==4.3.21
jedi==0.17.0
Kivy==1.11.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
parso==0.7.0
pycparser==2.20
pygame==2.0.0.dev8
pygame-sdl2==2.1.0
pyjnius==1.2.1
pylint==2.5.2
PyNaCl==1.4.0
PySDL2==0.9.7
pyTelegramBotAPI==4.0.0
python-telegram-bot==13.7
pytz==2021.1
requests==2.26.0
schedule==1.1.0
six==1.15.0
toml==0.10.1
tornado==6.1
typed-ast==1.4.1
tzlocal==3.0
urllib3==1.26.6
wrapt==1.12.1
